# docs/swagger-spec.json
{"openapi":"3.0.0","paths":{"/carriers":{"get":{"operationId":"CarrierController_get","parameters":[{"name":"page","required":true,"in":"query","description":"Requested page","schema":{"example":1,"type":"number"}},{"name":"limit","required":true,"in":"query","description":"Requested page limit","schema":{"example":10,"type":"number"}},{"name":"search","required":false,"in":"query","description":"Search keyword for carriers. Searches by model and license plate","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","description":"Desired status of carriers.","schema":{"type":"string","enum":["AVAILABLE","BUSY"]}}],"responses":{"200":{"description":"Returns a paginated list of carriers","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedCarrierResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get paginated carriers","tags":["Carriers"]},"post":{"operationId":"CarrierController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCarrierRequestDto"}}}},"responses":{"201":{"description":"Carrier created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrierResponseDto"}}}},"422":{"description":"Currency conversion error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrierUnprocessableEntityException"}}}}},"security":[{"bearer":[]}],"summary":"Crate a new carrier","tags":["Carriers"]}},"/carriers/{id}":{"get":{"operationId":"CarrierController_getCarrierById","parameters":[{"name":"id","required":true,"in":"path","description":"The UUID of record.","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a carrier by ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrierResponseDto"}}}},"404":{"description":"Carrier not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrierNotFoundException"}}}}},"security":[{"bearer":[]}],"summary":"Get carrier by ID","tags":["Carriers"]},"put":{"operationId":"CarrierController_update","parameters":[{"name":"id","required":true,"in":"path","description":"The UUID of record.","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCarrierBody"}}}},"responses":{"200":{"description":"Carrier updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrierResponseDto"}}}},"400":{"description":"Cannot edit carrier while fulfilling order","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrierBadRequestException"}}}},"404":{"description":"Carrier not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrierNotFoundException"}}}},"422":{"description":"Currency has to be defined in order to update rate","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrierUnprocessableEntityException"}}}}},"security":[{"bearer":[]}],"summary":"Create a carrier","tags":["Carriers"]},"delete":{"operationId":"CarrierController_delete","parameters":[{"name":"id","required":true,"in":"path","description":"The UUID of record.","schema":{"type":"string"}}],"responses":{"200":{"description":"Carrier deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrierResponseDto"}}}},"404":{"description":"Carrier not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrierNotFoundException"}}}}},"security":[{"bearer":[]}],"summary":"Delete carrier","tags":["Carriers"]}},"/routes":{"get":{"operationId":"RouteController_getAll","parameters":[{"name":"page","required":true,"in":"query","description":"Requested page","schema":{"example":1,"type":"number"}},{"name":"limit","required":true,"in":"query","description":"Requested page limit","schema":{"example":10,"type":"number"}},{"name":"status","required":false,"in":"query","description":"Desired status of carriers.","schema":{"type":"string","enum":["AWAITING_DISPATCH","IN_PROGRESS","COMPLETED"]}},{"name":"maxPrice","required":false,"in":"query","description":"Maximum price of the route","schema":{"type":"number"}},{"name":"minPrice","required":false,"in":"query","description":"Minimum price of the route","schema":{"type":"number"}}],"responses":{"200":{"description":"Returns a paginated list of routes","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetRoutesResponseDto"}}}},"400":{"description":"Minimum price can not be greater than maximum price","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PriceRangeException"}}}}},"security":[{"bearer":[]}],"summary":"Get paginated routes","tags":["Routes"]},"post":{"operationId":"RouteController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRouteRequestDto"}}}},"responses":{"201":{"description":"Route created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteResponseDto"}}}},"400":{"description":"The departure date can not be greater than completionDate or carrier is not available","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteBadRequestException"}}}},"500":{"description":"The distance between the coordinates can not be calculated"}},"security":[{"bearer":[]}],"summary":"Create a route","tags":["Routes"]}},"/routes/{id}":{"get":{"operationId":"RouteController_getOne","parameters":[{"name":"id","required":true,"in":"path","description":"The UUID of record.","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a route by ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteResponseDto"}}}},"404":{"description":"Route not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteNotFoundException"}}}}},"security":[{"bearer":[]}],"summary":"Find route by ID with assigned carrier if present","tags":["Routes"]},"put":{"operationId":"RouteController_update","parameters":[{"name":"id","required":true,"in":"path","description":"The UUID of record.","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRouteBody"}}}},"responses":{"200":{"description":"Route updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteResponseDto"}}}},"400":{"description":"Departure date can not be greater than completion date, carrier already assigned, or carrier not available","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteBadRequestException"}}}},"404":{"description":"Route not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteNotFoundException"}}}}},"security":[{"bearer":[]}],"summary":"Update route basic info","tags":["Routes"]},"delete":{"operationId":"RouteController_delete","parameters":[{"name":"id","required":true,"in":"path","description":"The UUID of record.","schema":{"type":"string"}}],"responses":{"200":{"description":"Route deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteResponseDto"}}}},"404":{"description":"Route not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteNotFoundException"}}}}},"security":[{"bearer":[]}],"summary":"Delete route by ID","tags":["Routes"]}},"/routes/{id}/carrier":{"put":{"description":"Sets carrier for the route","operationId":"RouteController_setCarrier","parameters":[{"name":"id","required":true,"in":"path","description":"The UUID of record.","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetRouteBody"}}}},"responses":{"200":{"description":"Route updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteResponseDto"}}}},"400":{"description":"Carrier has been already assigned to this route","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrierAlreadyAssignedException"}}}},"404":{"description":"Route not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteNotFoundException"}}}}},"security":[{"bearer":[]}],"summary":"Assign carrier to the route","tags":["Routes"]}},"/routes/{id}/status":{"put":{"operationId":"RouteController_updateStatus","parameters":[{"name":"id","required":true,"in":"path","description":"The UUID of record.","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRouteStatusBody"}}}},"responses":{"200":{"description":"Route updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteBadRequestException"}}}},"404":{"description":"Route not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteNotFoundException"}}}}},"security":[{"bearer":[]}],"summary":"Update route status","tags":["Routes"]}}},"info":{"title":"Freight-Tracker","description":"Freight-Tracker API documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CarrierResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier of the carrier","example":"123e4567-e89b-12d3-a456-426614174000"},"licensePlate":{"type":"string","description":"License plate of the carrier","example":"ABC123"},"model":{"type":"string","description":"Model of the carrier","example":"Volvo FH16"},"type":{"type":"string","description":"Type of the carrier","enum":["MINI","BOX","FLAT","REFRIGERATED","TANKER"],"example":"BOX"},"registrationDate":{"format":"date-time","type":"string","description":"Registration date of the carrier","example":"2023-01-01T00:00:00.000Z"},"status":{"type":"string","description":"Current status of the carrier","enum":["AVAILABLE","BUSY"],"example":"AVAILABLE"},"rate":{"type":"number","description":"Rate per kilometer in USD","example":2.5},"createdAt":{"format":"date-time","type":"string","description":"Creation timestamp","example":"2023-01-01T00:00:00.000Z"},"updatedAt":{"format":"date-time","type":"string","description":"Last update timestamp","example":"2023-01-01T00:00:00.000Z"}},"required":["id","licensePlate","model","type","registrationDate","status","rate","createdAt","updatedAt"]},"PaginatedCarrierResponseDto":{"type":"object","properties":{"data":{"description":"Array of carrier objects","type":"array","items":{"$ref":"#/components/schemas/CarrierResponseDto"}},"total":{"type":"number","description":"Total number of records available."}},"required":["data","total"]},"CarrierNotFoundException":{"type":"object","properties":{"statusCode":{"type":"number","example":404},"message":{"type":"string","example":"Carrier with specified ID has been not found"},"error":{"type":"string","example":"Not Found"}},"required":["statusCode","message","error"]},"CreateCarrierRequestDto":{"type":"object","properties":{"licensePlate":{"type":"string","description":"License plate number of the carrier","example":"ABC-1234"},"model":{"type":"string","description":"Model of the carrier vehicle","example":"Mercedes-Benz Sprinter"},"type":{"type":"string","description":"Type of carrier vehicle","enum":["MINI","BOX","FLAT","REFRIGERATED","TANKER"],"example":"TANKER"},"registrationDate":{"type":"string","description":"Purchase date of the carrier","example":"2024-01-15T00:00:00.000Z"},"rate":{"type":"number","description":"Price per kilometer","example":1.5},"currency":{"type":"string","description":"Currency of the rate","enum":["USD","EUR","UAH"],"example":"EUR"}},"required":["licensePlate","model","type","registrationDate","rate","currency"]},"CarrierUnprocessableEntityException":{"type":"object","properties":{"statusCode":{"type":"number","example":422},"message":{"type":"string","example":"Currency has to be defined in order to update rate"},"error":{"type":"string","example":"Unprocessable Entity"}},"required":["statusCode","message","error"]},"UpdateCarrierBody":{"type":"object","properties":{"licensePlate":{"type":"string","description":"License plate number of the carrier","example":"ABC-1234"},"model":{"type":"string","description":"Model of the carrier vehicle","example":"Mercedes-Benz Sprinter"},"type":{"type":"string","description":"Type of carrier vehicle","enum":["MINI","BOX","FLAT","REFRIGERATED","TANKER"],"example":"BOX"},"purchaseDate":{"type":"string","description":"Purchase date of the carrier","example":"2024-01-15T00:00:00.000Z"},"status":{"type":"string","description":"Status of the carrier","enum":["AVAILABLE","BUSY"],"example":"AVAILABLE"},"rate":{"type":"number","description":"Price per kilometer","example":1.5},"currency":{"type":"string","description":"Currency of the rate","enum":["USD","EUR","UAH"],"example":"EUR"}}},"CarrierBadRequestException":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"message":{"type":"string","example":"Can not edit carrier while fulfilling order"},"error":{"type":"string","example":"Bad Request"}},"required":["statusCode","message","error"]},"RouteResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier of the route","example":"123e4567-e89b-12d3-a456-426614174000"},"startPoint":{"type":"string","description":"Starting point coordinates","example":"52.2297,21.0122"},"endPoint":{"type":"string","description":"End point coordinates","example":"51.5074,0.1278"},"distance":{"type":"number","description":"Distance in meters","example":1500.75},"departureDate":{"format":"date-time","type":"string","description":"Planned departure date","example":"2023-01-01T00:00:00.000Z"},"completionDate":{"format":"date-time","type":"string","description":"Planned completion date","example":"2023-01-02T00:00:00.000Z"},"departureDateActual":{"format":"date-time","type":"string","description":"Actual departure date","example":"2023-01-01T01:30:00.000Z"},"completionDateActual":{"format":"date-time","type":"string","description":"Actual completion date","example":"2023-01-02T02:15:00.000Z"},"requiredCarrierType":{"type":"string","description":"Required carrier type for this route","enum":["MINI","BOX","FLAT","REFRIGERATED","TANKER"],"example":"BOX"},"carrierFee":{"type":"number","description":"Fee paid to carrier in USD","example":250.5},"price":{"type":"number","description":"Price of the route in USD","example":500.75},"status":{"type":"string","description":"Current status of the route","enum":["AWAITING_DISPATCH","IN_PROGRESS","COMPLETED"],"example":"AWAITING_DISPATCH"},"carrierId":{"type":"string","description":"ID of the assigned carrier","example":"123e4567-e89b-12d3-a456-426614174000"},"carrier":{"description":"Assigned carrier details","allOf":[{"$ref":"#/components/schemas/CarrierResponseDto"}]},"createdAt":{"format":"date-time","type":"string","description":"Creation timestamp","example":"2023-01-01T00:00:00.000Z"},"updatedAt":{"format":"date-time","type":"string","description":"Last update timestamp","example":"2023-01-01T00:00:00.000Z"}},"required":["id","startPoint","endPoint","distance","requiredCarrierType","price","status","createdAt","updatedAt"]},"GetRoutesResponseDto":{"type":"object","properties":{"data":{"description":"Array of route objects","type":"array","items":{"$ref":"#/components/schemas/RouteResponseDto"}},"total":{"type":"number","description":"Total number of records available."}},"required":["data","total"]},"PriceRangeException":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"message":{"type":"string","example":"Minimum price can not be greater than maximum price"},"error":{"type":"string","example":"Bad Request"}},"required":["statusCode","message","error"]},"RouteNotFoundException":{"type":"object","properties":{"statusCode":{"type":"number","example":404},"message":{"type":"string","example":"Route with provided ID has been not found."},"error":{"type":"string","example":"Not Found"}},"required":["statusCode","message","error"]},"CreateRouteRequestDto":{"type":"object","properties":{"startPoint":{"type":"string","description":"The coordinates of the start city","example":"50.4504,30.5245"},"endPoint":{"type":"string","description":"The coordinates of the end city","example":"52.5200,13.4050"},"departureDate":{"type":"string","description":"Scheduled departure date and time","example":"2024-12-01T08:00:00.000Z"},"completionDate":{"type":"string","description":"Actual completion date and time","example":"2024-12-01T16:30:00.000Z"},"requiredCarrierType":{"type":"string","description":"Type of carrier required for this route","enum":["MINI","BOX","FLAT","REFRIGERATED","TANKER"],"example":"BOX"},"price":{"type":"number","description":"Price for the route","example":450},"currency":{"type":"string","description":"Currency of the price","enum":["USD","EUR","UAH"],"example":"EUR"}},"required":["startPoint","endPoint","departureDate","completionDate","requiredCarrierType","price","currency"]},"RouteBadRequestException":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"message":{"type":"string","example":"The departure date can not be greater than completionDate"},"error":{"type":"string","example":"Bad Request"}},"required":["statusCode","message","error"]},"UpdateRouteBody":{"type":"object","properties":{"departureDate":{"type":"string","description":"Scheduled departure date and time","example":"2024-12-01T08:00:00.000Z"},"completionDate":{"type":"string","description":"Actual completion date and time","example":"2024-12-01T16:30:00.000Z"},"requiredCarrierType":{"type":"string","description":"Type of carrier required for this route","enum":["MINI","BOX","FLAT","REFRIGERATED","TANKER"],"example":"BOX"},"price":{"type":"number","description":"Price for the route","example":450},"currency":{"type":"string","description":"Currency of the price","enum":["USD","EUR","UAH"],"example":"EUR"}}},"SetRouteBody":{"type":"object","properties":{"carrierId":{"type":"string","description":"ID of the assigned carrier","example":"d899f5ac-228f-4974-99ac-0466c0de2e93"}}},"CarrierAlreadyAssignedException":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"message":{"type":"string","example":"Carrier has been already assigned to this route"},"error":{"type":"string","example":"Bad Request"}},"required":["statusCode","message","error"]},"UpdateRouteStatusBody":{"type":"object","properties":{"departureDateActual":{"type":"string","description":"Scheduled departure date and time","example":"2024-12-01T08:00:00.000Z"},"completionDateActual":{"type":"string","description":"Actual completion date and time","example":"2024-12-01T16:30:00.000Z"},"status":{"type":"string","description":"Status of the route","example":"IN_PROGRESS","enum":["AWAITING_DISPATCH","IN_PROGRESS","COMPLETED"]}}}}}}