# src/api/domain/carrier/dto/response/paginated-carrier.response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { PaginatedResponseDto } from '../../../../shared';
import { CarrierResponseDto } from './carrier.response.dto';

export class PaginatedCarrierResponseDto extends PaginatedResponseDto<CarrierResponseDto> {
  @ApiProperty({
    description: 'Array of carrier objects',
    type: [CarrierResponseDto],
  })
  data: CarrierResponseDto[];
}